[Important] Fixed bug in rpc_utils.altitude_range function.
We were adding two 1D arrays of the same length (srtm and geoid), but
one of the two (srtm) was stored as a 2D array with one column. Oddly if
you add a 1D array of length 5 with a 2D array of dimensions (5, 1), in
numpy, you end up with a 2D array of dimensions (5, 5)... This 2D array
contains all the possible sums between the coefficients of the two input
vectors. Thus its smallest element is smaller than the smallest element
of the sum of the two vectors.

This bug may have had two consequences:
* the estimated altitude ranges were too wide
* for big areas, the running time was longer