[IMPORTANT] Fixed bug in srtm4 binary.
An end of line character, "\n", was missing in a sscanf call. As a
consequence there was a shift of one sample: the srtm sample that was
read in the ASCII file was the one just before the right one. As the
srtm is rather smooth, it is not that bad, but in some cases it may have
been critical: for example when the sample to be read was (0, 0), the
bugged code was returning the atoi value of "\n", ie -35.