improved handling of tiles where no sift matches are found:
* if no sift matches are found for a given tile, the file
  sift_matches.txt is not written
* if no pointing correction can be computed from a given tile, the file
  pointing.txt is not written

Tiles for which the file pointing.txt is missing are thus reprocessed
sequentially,
using a pointing correction matrix estimated from the neighboring tiles
(if available), or from the global correction matrix (if available), or
identity matrix. At the moment the function
pointing_accuray.from_next_tiles is not implemented and always returns
identity.