added new config boolean parameter 'fusion_conservative'
Only useful for triplets. When set to True, the height map fusion is
more conservative, in the sense that if a pixel does not have a valid
height in one of the two height maps then it is rejected. Only the
pixels whose two heights are valid and agree up to the given threshold
are kept.

An obvious drawback of this conservative strategy is that points that
are visible in only two of the three views will be discarded.