replaced rand() (from stdlib) in RANSAC to fix the seed.
The stdlib rand() function is replaced with mt, a very slightly amended
version of the fast, portable and high-quality Mersenne Twister
pseudo-random number mt19937ar by Makoto Matsumoto and Takuji Nishimura.

By imposing the seed we are able to obtain the same results one the same
inputs when running on different machines.