re-implemented the intra-n-plet fusion algorithm as it used to be
Let n be the number of images. When n is greater or equal to 3, the n-1
height maps obtained from the pairs that share a common reference image
are merged pixelwise (with the median, the mean or the min value of the
n-1 samples). Before merging, however, the height maps need an altimetry
registration, ie the height maps need to be shifted in the intensity
space in order to have all the same mean. This registration has to be
done on the whole height map, covering the whole ROI being processed by
s2p. If done tilewise, it will create visible border artifacts on the
tiles borders.