remove overlap between tiles
The only reason why we were using overlapping tiles is that the
disparity estimation works poorly on the left and right borders of
rectified images, because there is less context and because the
corresponding point may be out of the secondary image footprint.
To fix this it's enough to extend horizontally the rectified images.
The "extra disparities" are discarded anyway during the triangulation step.