Remove colorization stuff.
s2p is a tool for computing elevation models. It's not a tool for
computing pansharpening and retinex. There are much better tools for
these tasks. The tentative implementation of pansharpening and retinex
that we have in s2p makes the whole pipeline more complicated and
contributes to its unmantainability.