Completed removal of colorization stuff.
If present, the color ('clr') version of the reference image is supposed
to be sampled on the same grid as the reference image (hence same
resolution) and to be encoded on 8 bits. It can be either graylevel or
rgb.

If not present, the point cloud of each tile is colorized with a linear
requantization of the reference image crop on [0, 255]. This will
produce visible tilewise color artifacts on the full point cloud. To
avoid such artifacts the user must prepare a full resolution 8 bits
version of the reference image (and optionally colorize it).