(bug) Fix relative paths inconsistencies
This PR tries to solve the following issue: when relative path is used as out_dir, the config files written in out_dir can only be used from the same cwd as the first s2p that generated the output dir.

As a result, the following will fail:

$ s2p.py config.json
$ cd somewhere-else
$ s2p.py path_to_output_dir/config.json

One solution would be to make all paths absolute in the config.json exported in output directory, but then we will loose the ability to move the output folder and still be able to use it.

This PR proposes the following solution:
1) If a relative path is used for out_dir in config.json, then s2p.py interprets it as relative to the config.json path, and not the cwd. If out_dir is absolute, it is left unchanged.

2) config.json exported in output dir always contain relative path : $out_dir/config.json has "." as out_dir and $out_dir/row_*/col_*/config.json has "../.." as outdir. One can therefore move the outdir and still get a working config.json.

3) All input data path are made absolute at config.json loading, and $out_dir/config.json therefore contain absolute path for all input data.

A side effect is that to allow for 1) we need to load the config.json through a dedicated helper function, and not through json.load(). This is because we need the path to the json file to update the out_dir in the dictionnary.