Use shell=False as default in common.run
shell=True was causing problems with the timeout argument, because the
children processes of the shell process were not being killed
(see https://stackoverflow.com/a/4789957/9977650).
Also, subprocess.check_call was replaced with the higher-level
subprocess.run.
All commands that used pipes or double quotes to escape spaces had to be
modified accordingly.