Introduce s2p's cfg["images"][?]["rpcm"] key
This key is used to remove the ambiguous nature of the "rpc" key,
which was used both by the user to provide the path to a custom RPC
file, and internally to store the RPCModel instance associated to each
image.
The "rpcm" key now stores the RPCModel instances, and the "rpc" key
has the same use as before.
Thanks to release 1.4.6 of rpcm, it is now also possible to provide
directly a dict with the RPC values in the "rpc" key.