Fix use of mgm_timeout config option
Python has an early binding mechanism for default arguments. They are
evaluated at a function's definition time, not at call time, so the
value of cfg['mgm_timeout'] in the definition of compute_disparity_map
was not the value potentially overriden by the user in his config.
See https://stackoverflow.com/a/1651284/9977650