Change initialization of rpc.c localization
This commit changes the default points used to initialize the recursive
localization function. The old points were the corners of the RPC
validity domain, and the new points are closer to the center of the
validity domain.

This was done to fix erratic behavior found for RPCs that have
unreasonably large validity domains (such as Skysat L1A), for which
initializing to the corners of the domain would not always converge.

See similar change in rpcm: https://github.com/cmla/rpcm/pull/11