Pin to pyproj version <3.0.0
In pyproj 3, wheels no longer include datumgrids because they are
accessible over the network (see
https://pyproj4.github.io/pyproj/stable/transformation_grids.html).
When not explicitly told to do so, pyproj 3 does an "approximate"
transformation, which changes the behavior of the
s2p.geographiclib.geoid_to_ellipsoid function and breaks the tests.
We are pinning pyproj<3.0.0 until a robust way of remotely accessing the
datumgrids is found.